requirements.txt
typer
rich
sqlalchemy
openai
python-dotenv
google-genai
python-dotenv